add a counter to the show on the header
