+++ src/index.html
@@ -16,7 +16,7 @@
             <a href="#"><img src="" id="logo" alt="logo"></a>
           </li>
           <li class="list-item">
-            <a href="#">Shows(8)</a>
+            <a href="#">Shows(<span class="shows-counter"></span>)</a>
           </li>
           <li class="list-item">
             <a href="#">Episodes</a>
